requirements.txt
Flask
flask-cors
gunicorn
numpy
PIllow
https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp36-cp36m-linux_x86_64.whl